requirements.txt
pyyaml
tqdm
attrdict
h5py
numpy
scipy
matplotlib
seaborn
brewer2mpl
torch==1.3.1
torchvision==0.4.2
tensorboard
tensorboardX
opencv-python==4.0.1.23
opencv-contrib-python==4.0.1.24
scikit-learn
scikit-image
kornia==0.2.0
shapely
jupyter
bresenham
pyvista
omegaconf
mathutils
trimesh
rtree
plyfile
pathlib
./third-party/pytlsd
